+++ CHANGELOG.md
@@ -158,6 +158,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Feature
 - Added hotkey for query window in data research v2
+- Added pagination on interesting fields
 
 ### Fixed
 - fixed data exploration 2.0 gantt: the date format setting for the Gantt chart does not apply