+++ CHANGELOG.md
@@ -8,8 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added columnCount, tileStyle options on dashTile
+- Added event handling on dashTile
 
 ### Fixed
+- Fixed error messages fo password un profile
+- Fix yGraph moving issue
 - fixed reportV2 height visualization
 
 ## [1.97.0]
@@ -157,6 +161,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Feature
 - Added hotkey for query window in data research v2
+- Added pagination on interesting fields
 
 ### Fixed
 - fixed data exploration 2.0 gantt: the date format setting for the Gantt chart does not apply