+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed error messages fo password un profile
 - Fix yGraph moving issue
+- fixed reportV2 height visualization
 
 ## [1.97.0]