+++ CHANGELOG.md
@@ -8,10 +8,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added columnCount, tileStyle options on dashTile
+- Added event handling on dashTile
 
 - Added Constructor-schemes component
 
 ### Fixed
+- Fixed error messages fo password un profile
+- Fix yGraph moving issue
+- fixed reportV2 height visualization
+- fix filter datapicker
+- Fixing versions of third-party components
 
 ## [1.97.0]